+++ report.php
@@ -242,6 +242,7 @@ class quiz_archiver_report extends report_base {
      * @param bool $exportquizbackup Complete quiz backup will be archived if true
      * @param bool $exportcoursebackup Complete course backup will be archived if true
      * @param string $archivefilenamepattern Filename pattern to use for archive generation
+     * @param string $attemptsfoldernamepattern Folder name pattern to use for attempt report generation
      * @param string $attemptsfilenamepattern Filename pattern to use for attempt report generation
      * @param array|null $imageoptimize If set, images in the attempt report will be optimized
      * according to the passed array containing 'width', 'height', and 'quality'
@@ -261,6 +262,7 @@ class quiz_archiver_report extends report_base {
         bool   $exportquizbackup,
         bool   $exportcoursebackup,
         string $archivefilenamepattern,
+        string $attemptsfoldernamepattern,
         string $attemptsfilenamepattern,
         ?array $imageoptimize = null,
         ?int   $retentionseconds = null
@@ -311,6 +313,7 @@ class quiz_archiver_report extends report_base {
                 'sections' => $reportsections,
                 'paper_format' => $paperformat,
                 'keep_html_files' => $reportkeephtmlfiles,
+                'foldername_pattern' => $attemptsfoldernamepattern,
                 'filename_pattern' => $attemptsfilenamepattern,
                 'image_optimize' => $imageoptimize ?? false,
             ];
@@ -540,6 +543,7 @@ class quiz_archiver_report extends report_base {
                         $formdata->export_quiz_backup,
                         $formdata->export_course_backup,
                         $formdata->archive_filename_pattern,
+                        $formdata->export_attempts_foldername_pattern,
                         $formdata->export_attempts_filename_pattern,
                         $formdata->export_attempts_image_optimize ? [
                             'width' => (int) $formdata->export_attempts_image_optimize_width,